64-bit sha1sum
When calculating the sha1sum of a block of memory, the size of that block was
limited to 32 bit. This patch changes it to 64-bit (on 64-bit platforms).

When later we add support for large harddisk images, we *might* actually have
to compute such large sha1sums. (Actually we'll likely incrementally compute
the sha1sum on many small subblocks, but there's no good reason to already
limit the possibility of using a large block at this level).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13080 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/utils/sha1.hh
@@ -58,14 +58,14 @@ public:
 	SHA1();
 
 	/** Incrementally calculate the hash value. */
-	void update(const byte* data, unsigned len);
+	void update(const byte* data, size_t len);
 
 	/** Get the final hash. After this method is called, calls to update()
 	  * are invalid. */
 	Sha1Sum digest();
 
 	/** Easier to use interface, if you can pass all data in one go. */
-	static Sha1Sum calc(const byte* data, unsigned len);
+	static Sha1Sum calc(const byte* data, size_t len);
 
 	/** Easier to use interface, if you can pass all data in one go. But
 	  * also report progress.
@@ -73,7 +73,7 @@ public:
 	  * completely, in one call. The caller is responsible to make sure
 	  * this is the case.
 	  */
-	static Sha1Sum calcWithProgress(const byte* data, unsigned len, const
+	static Sha1Sum calcWithProgress(const byte* data, size_t len, const
 			std::string& filename, CliComm& cliComm,
 			EventDistributor& distributor);
